Separate the CloudWatch module from MetricsSender
